python/requirements.txt
torch
pytorch-geometric
numpy
h5py
matplotlib
pybullet
mujoco
wandb
